esm - reduce diff
