+++ src/bootstrap-amd.js
@@ -19,17 +19,33 @@ const isESM = false;
 // import * as fs from 'fs';
 // import { fileURLToPath } from 'url';
 // import { createRequire, register } from 'node:module';
-// if (process.env['ELECTRON_RUN_AS_NODE'] || process.versions['electron']) {
-// 	register('./loader-original-fs.mjs', import.meta.url);
-// }
 // import { product, pkg } from './bootstrap-meta.js';
-// import * as bootstrap from './bootstrap.js';
+// import * as bootstrapNode from './bootstrap-node.js';
 // import * as performance from './vs/base/common/performance.js';
 //
 // const require = createRequire(import.meta.url);
 // const isESM = true;
 // const module = { exports: {} };
 // const __dirname = path.dirname(fileURLToPath(import.meta.url));
+//
+// // Install a hook to module resolution to map 'fs' to 'original-fs'
+// if (process.env['ELECTRON_RUN_AS_NODE'] || process.versions['electron']) {
+// 	const jsCode = `
+// 	export async function resolve(specifier, context, nextResolve) {
+// 		if (specifier === 'fs') {
+// 			return {
+// 				format: 'builtin',
+// 				shortCircuit: true,
+// 				url: 'node:original-fs'
+// 			};
+// 		}
+
+// 		// Defer to the next hook in the chain, which would be the
+// 		// Node.js default resolve if this is the last user-specified loader.
+// 		return nextResolve(specifier, context);
+// 	}`;
+// 	register(`data:text/javascript;base64,${Buffer.from(jsCode).toString('base64')}`, import.meta.url);
+// }
 // ESM-uncomment-end
 
 // Store the node.js require function in a variable
@@ -67,7 +83,7 @@ globalThis._VSCODE_PACKAGE_JSON = require('./bootstrap-meta').pkg;
 globalThis._VSCODE_FILE_ROOT = __dirname;
 
 // ESM-comment-begin
-const bootstrap = require('./bootstrap');
+const bootstrapNode = require('./bootstrap-node');
 const performance = require(`./vs/base/common/performance`);
 const fs = require('fs');
 // ESM-comment-end
@@ -109,7 +125,6 @@ async function doSetupNLS() {
 				messagesFile = nlsConfig.defaultMessagesFile;
 			}
 
-			// VSCODE_GLOBALS: NLS
 			globalThis._VSCODE_NLS_LANGUAGE = nlsConfig?.resolvedLanguage;
 		} catch (e) {
 			console.error(`Error reading VSCODE_NLS_CONFIG from environment: ${e}`);
@@ -124,7 +139,6 @@ async function doSetupNLS() {
 	}
 
 	try {
-		// VSCODE_GLOBALS: NLS
 		globalThis._VSCODE_NLS_MESSAGES = JSON.parse((await fs.promises.readFile(messagesFile)).toString());
 	} catch (error) {
 		console.error(`Error reading NLS messages file ${messagesFile}: ${error}`);
@@ -141,7 +155,6 @@ async function doSetupNLS() {
 		// Fallback to the default message file to ensure english translation at least
 		if (nlsConfig?.defaultMessagesFile && nlsConfig.defaultMessagesFile !== messagesFile) {
 			try {
-				// VSCODE_GLOBALS: NLS
 				globalThis._VSCODE_NLS_MESSAGES = JSON.parse((await fs.promises.readFile(nlsConfig.defaultMessagesFile)).toString());
 			} catch (error) {
 				console.error(`Error reading default NLS messages file ${nlsConfig.defaultMessagesFile}: ${error}`);
@@ -186,7 +199,7 @@ if (isESM) {
 	const loader = require('./vs/loader');
 
 	loader.config({
-		baseUrl: bootstrap.fileUriFromPath(__dirname, { isWindows: process.platform === 'win32' }),
+		baseUrl: bootstrapNode.fileUriFromPath(__dirname, { isWindows: process.platform === 'win32' }),
 		catchError: true,
 		nodeRequire,
 		amdModulesPattern: /^vs\//,
