+++ src/bootstrap-node.js
@@ -9,6 +9,7 @@
 // ESM-comment-begin
 const path = require('path');
 const fs = require('fs');
+const Module = require('module');
 
 const isESM = false;
 // ESM-comment-end
@@ -19,10 +20,29 @@ const isESM = false;
 // import { createRequire } from 'node:module';
 //
 // const require = createRequire(import.meta.url);
+// const Module = require('module');
 // const isESM = true;
 // const module = { exports: {} };
 // const __dirname = path.dirname(fileURLToPath(import.meta.url));
 // ESM-uncomment-end
+
+// increase number of stack frames(from 10, https://github.com/v8/v8/wiki/Stack-Trace-API)
+Error.stackTraceLimit = 100;
+
+if (!process.env['VSCODE_HANDLES_SIGPIPE']) {
+	// Workaround for Electron not installing a handler to ignore SIGPIPE
+	// (https://github.com/electron/electron/issues/13254)
+	let didLogAboutSIGPIPE = false;
+	process.on('SIGPIPE', () => {
+		// See https://github.com/microsoft/vscode-remote-release/issues/6543
+		// In certain situations, the console itself can be in a broken pipe state
+		// so logging SIGPIPE to the console will cause an infinite async loop
+		if (!didLogAboutSIGPIPE) {
+			didLogAboutSIGPIPE = true;
+			console.error(new Error(`Unexpected SIGPIPE`));
+		}
+	});
+}
 
 // Setup current working directory in all our node & electron processes
 // - Windows: call `process.chdir()` to always set application folder as cwd
@@ -177,8 +197,74 @@ module.exports.configurePortable = function (product) {
 	};
 };
 
+/**
+ * Helper to enable ASAR support.
+ */
+module.exports.enableASARSupport = function () {
+	const NODE_MODULES_PATH = path.join(__dirname, '../node_modules');
+	const NODE_MODULES_ASAR_PATH = `${NODE_MODULES_PATH}.asar`;
+
+	// @ts-ignore
+	const originalResolveLookupPaths = Module._resolveLookupPaths;
+
+	// @ts-ignore
+	Module._resolveLookupPaths = function (request, parent) {
+		const paths = originalResolveLookupPaths(request, parent);
+		if (Array.isArray(paths)) {
+			for (let i = 0, len = paths.length; i < len; i++) {
+				if (paths[i] === NODE_MODULES_PATH) {
+					paths.splice(i, 0, NODE_MODULES_ASAR_PATH);
+					break;
+				}
+			}
+		}
+
+		return paths;
+	};
+};
+
+/**
+ * Helper to convert a file path to a URI.
+ *
+ * TODO@bpasero check for removal once ESM has landed.
+ *
+ * @param {string} path
+ * @param {{ isWindows?: boolean, scheme?: string, fallbackAuthority?: string }} config
+ * @returns {string}
+ */
+module.exports.fileUriFromPath = function (path, config) {
+
+	// Since we are building a URI, we normalize any backslash
+	// to slashes and we ensure that the path begins with a '/'.
+	let pathName = path.replace(/\\/g, '/');
+	if (pathName.length > 0 && pathName.charAt(0) !== '/') {
+		pathName = `/${pathName}`;
+	}
+
+	/** @type {string} */
+	let uri;
+
+	// Windows: in order to support UNC paths (which start with '//')
+	// that have their own authority, we do not use the provided authority
+	// but rather preserve it.
+	if (config.isWindows && pathName.startsWith('//')) {
+		uri = encodeURI(`${config.scheme || 'file'}:${pathName}`);
+	}
+
+	// Otherwise we optionally add the provided authority if specified
+	else {
+		uri = encodeURI(`${config.scheme || 'file'}://${config.fallbackAuthority || ''}${pathName}`);
+	}
+
+	return uri.replace(/#/g, '%23');
+};
+
+//#endregion
+
 // ESM-uncomment-begin
 // export const injectNodeModuleLookupPath = module.exports.injectNodeModuleLookupPath;
 // export const removeGlobalNodeModuleLookupPaths = module.exports.removeGlobalNodeModuleLookupPaths;
 // export const configurePortable = module.exports.configurePortable;
+// export const enableASARSupport = module.exports.enableASARSupport;
+// export const fileUriFromPath = module.exports.fileUriFromPath;
 // ESM-uncomment-end
