+++ src/vs/base/common/platform.ts
@@ -103,7 +103,6 @@ else if (typeof navigator === 'object' && !isElectronRenderer) {
 	_isLinux = _userAgent.indexOf('Linux') >= 0;
 	_isMobile = _userAgent?.indexOf('Mobi') >= 0;
 	_isWeb = true;
-	// VSCODE_GLOBALS: NLS
 	_language = globalThis._VSCODE_NLS_LANGUAGE || LANGUAGE_DEFAULT;
 	_locale = navigator.language.toLowerCase();
 	_platformLocale = _locale;
