+++ src/vs/code/electron-sandbox/processExplorer/processExplorer-dev.html
@@ -35,8 +35,7 @@
 	</body>
 
 	<!-- Startup (do not modify order of script tags!) -->
-	<script src="../../../../bootstrap.js"></script>
 	<script src="../../../../vs/loader.js"></script>
 	<script src="../../../../bootstrap-window.js"></script>
-	<script src="processExplorer.js"></script>
+	<script src="./processExplorer.js"></script>
 </html>
