+++ src/vs/code/electron-sandbox/workbench/workbench-dev.html
@@ -68,8 +68,7 @@
 	</body>
 
 	<!-- Startup (do not modify order of script tags!) -->
-	<script src="../../../../bootstrap.js"></script>
 	<script src="../../../../vs/loader.js"></script>
 	<script src="../../../../bootstrap-window.js"></script>
-	<script src="workbench.js"></script>
+	<script src="./workbench.js"></script>
 </html>
