+++ src/vs/nls.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-// VSCODE_GLOBALS: NLS
 const isPseudo = globalThis._VSCODE_NLS_LANGUAGE === 'pseudo' || (typeof document !== 'undefined' && document.location && document.location.hash.indexOf('pseudo=true') >= 0);
 
 export interface ILocalizeInfo {
@@ -87,7 +86,6 @@ export function localize(data: ILocalizeInfo | string /* | number when built */,
  * depending on the target context.
  */
 function lookupMessage(index: number, fallback: string | null): string {
-	// VSCODE_GLOBALS: NLS
 	const message = globalThis._VSCODE_NLS_MESSAGES?.[index];
 	if (typeof message !== 'string') {
 		if (typeof fallback === 'string') {
