+++ src/vs/platform/environment/test/node/nativeModules.integrationTest.ts
@@ -118,21 +118,10 @@ flakySuite('Native Modules (all platforms)', () => {
 		assert.ok(typeof result === 'string' || typeof result === 'undefined', testErrorMessage('@vscode/windows-registry'));
 	});
 
-	test('@vscode/windows-ca-certs', async () => {
-		// @ts-ignore we do not directly depend on this module anymore
-		// but indirectly from our dependency to `@vscode/proxy-agent`
-		// we still want to ensure this module can work properly.
-		const windowsCerts = await import('@vscode/windows-ca-certs');
-		const store = new windowsCerts.Crypt32();
-		assert.ok(windowsCerts, testErrorMessage('@vscode/windows-ca-certs'));
-		let certCount = 0;
-		try {
-			while (store.next()) {
-				certCount++;
-			}
-		} finally {
-			store.done();
-		}
-		assert(certCount > 0);
+	test('@vscode/proxy-agent', async () => {
+		const proxyAgent = await import('@vscode/proxy-agent');
+		// This call will load `@vscode/proxy-agent` which is a native module that we want to test on Windows
+		const windowsCerts = await proxyAgent.loadSystemCertificates({ log: console });
+		assert.ok(windowsCerts.length > 0, testErrorMessage('@vscode/proxy-agent'));
 	});
 });
