+++ src/vs/platform/issue/electron-main/issueMainService.ts
@@ -83,7 +83,6 @@ export class IssueMainService implements IIssueMainService {
 					},
 					product,
 					nls: {
-						// VSCODE_GLOBALS: NLS
 						messages: globalThis._VSCODE_NLS_MESSAGES,
 						language: globalThis._VSCODE_NLS_LANGUAGE
 					}
