+++ src/vs/server/node/remoteExtensionHostAgentServer.ts
@@ -696,7 +696,7 @@ export async function createServer(address: string | net.AddressInfo | null, arg
 	let didLogAboutSIGPIPE = false;
 	process.on('SIGPIPE', () => {
 		// See https://github.com/microsoft/vscode-remote-release/issues/6543
-		// We would normally install a SIGPIPE listener in bootstrap.js
+		// We would normally install a SIGPIPE listener in bootstrap-node.js
 		// But in certain situations, the console itself can be in a broken pipe state
 		// so logging SIGPIPE to the console will cause an infinite async loop
 		if (!didLogAboutSIGPIPE) {
