+++ src/vs/workbench/contrib/issue/electron-sandbox/issueReporter-dev.html
@@ -39,8 +39,7 @@
 	</body>
 
 	<!-- Startup (do not modify order of script tags!) -->
-	<script src="../../../../../bootstrap.js"></script>
 	<script src="../../../../../vs/loader.js"></script>
 	<script src="../../../../../bootstrap-window.js"></script>
-	<script src="issueReporter.js"></script>
+	<script src="./issueReporter.js"></script>
 </html>
