+++ src/vs/workbench/contrib/issue/electron-sandbox/issueReporter.html
@@ -39,5 +39,5 @@
 	</body>
 
 	<!-- Startup (do not modify order of script tags!) -->
-	<script src="issueReporter.js"></script>
+	<script src="./issueReporter.js"></script>
 </html>
