+++ src/vs/workbench/services/extensions/browser/webWorkerExtensionHost.ts
@@ -6,6 +6,7 @@
 import * as dom from 'vs/base/browser/dom';
 import { parentOriginHash } from 'vs/base/browser/iframe';
 import { mainWindow } from 'vs/base/browser/window';
+import { isESM } from 'vs/base/common/amd';
 import { Barrier } from 'vs/base/common/async';
 import { VSBuffer } from 'vs/base/common/buffer';
 import { canceled, onUnexpectedError } from 'vs/base/common/errors';
@@ -183,14 +184,14 @@ export class WebWorkerExtensionHost extends Disposable implements IExtensionHost
 			}
 			if (event.data.type === 'vscode.bootstrap.nls') {
 				const factoryModuleId = 'vs/base/worker/workerMain.js';
-				const baseUrl = require.toUrl(factoryModuleId).slice(0, -factoryModuleId.length);
+				const baseUrl = isESM ? undefined : require.toUrl(factoryModuleId).slice(0, -factoryModuleId.length);
 				iframe.contentWindow!.postMessage({
 					type: event.data.type,
 					data: {
 						baseUrl,
-						workerUrl: require.toUrl(factoryModuleId),
+						workerUrl: isESM ? FileAccess.asBrowserUri(factoryModuleId).toString(true) : require.toUrl(factoryModuleId),
+						fileRoot: globalThis._VSCODE_FILE_ROOT,
 						nls: {
-							// VSCODE_GLOBALS: NLS
 							messages: globalThis._VSCODE_NLS_MESSAGES,
 							language: globalThis._VSCODE_NLS_LANGUAGE
 						}
