+++ src/vs/workbench/services/extensions/worker/webWorkerExtensionHostIframe.html
@@ -4,7 +4,7 @@
 		<meta http-equiv="Content-Security-Policy" content="
 			default-src 'none';
 			child-src 'self' data: blob:;
-			script-src 'self' 'unsafe-eval' 'sha256-V28GQnL3aYxbwgpV3yW1oJ+VKKe/PBSzWntNyH8zVXA=' https: http://localhost:* blob:;
+			script-src 'self' 'unsafe-eval' 'sha256-2Q+j4hfT09+1+imS46J2YlkCtHWQt0/BE79PXjJ0ZJ8=' https: http://localhost:* blob:;
 			connect-src 'self' https: wss: http://localhost:* http://127.0.0.1:* ws://localhost:* ws://127.0.0.1:*;"/>
 	</head>
 	<body>
@@ -78,7 +78,7 @@
 				return;
 			}
 			const { data } = event.data;
-			createWorker(data.baseUrl, data.workerUrl, data.nls.messages, data.nls.language);
+			createWorker(data.baseUrl, data.workerUrl, data.fileRoot, data.nls.messages, data.nls.language);
 		};
 
 		window.parent.postMessage({
@@ -87,24 +87,35 @@
 		}, '*');
 	}
 
-	function createWorker(baseUrl, workerUrl, nlsMessages, nlsLanguage) {
+	function createWorker(baseUrl, workerUrl, fileRoot, nlsMessages, nlsLanguage) {
 		try {
 			if (globalThis.crossOriginIsolated) {
 				workerUrl += '?vscode-coi=2'; // COEP
 			}
 
+			// ESM-comment-begin
+			const isESM = false;
+			// ESM-comment-end
+			// ESM-uncomment-begin
+			// const isESM = true;
+			// ESM-uncomment-end
+
 			const blob = new Blob([[
 				`/*extensionHostWorker*/`,
 				`globalThis.MonacoEnvironment = { baseUrl: '${baseUrl}' };`,
-				// VSCODE_GLOBALS: NLS
 				`globalThis._VSCODE_NLS_MESSAGES = ${JSON.stringify(nlsMessages)};`,
 				`globalThis._VSCODE_NLS_LANGUAGE = ${JSON.stringify(nlsLanguage)};`,
-				`importScripts('${workerUrl}');`,
+				`globalThis._VSCODE_FILE_ROOT = '${fileRoot}';`,
+				isESM ? `await import('${workerUrl}');` : `importScripts('${workerUrl}');`,
+				isESM ? `globalThis.onmessage({ data: 'vs/workbench/api/worker/extensionHostWorker' });` : undefined, // important to start loading after the ESM async import has finished
 				`/*extensionHostWorker*/`
 			].join('')], { type: 'application/javascript' });
 
-			const worker = new Worker(URL.createObjectURL(blob), { name });
-			worker.postMessage('vs/workbench/api/worker/extensionHostWorker');
+			const worker = new Worker(URL.createObjectURL(blob), { name, type: isESM ? 'module' : undefined });
+			if (!isESM) {
+				// Note: cannot postMessage into a worker that is ESM because imports are async
+				worker.postMessage('vs/workbench/api/worker/extensionHostWorker');
+			}
 			const nestedWorkers = new Map();
 
 			worker.onmessage = (event) => {
